build: move certificate creation and deletion to respective build files
