+++ drone-system/build.gradle.kts
@@ -1,3 +1,5 @@
+import java.util.Properties
+
 plugins {
     id("dronesecurity-application")
     id("de.jjohannes.extra-java-module-info")
@@ -25,6 +27,64 @@ extraJavaModuleInfo {
 }
 
 tasks {
+    val propertiesFile = file(
+            (if (System.getenv("CI") == true.toString())
+                projectDir
+            else
+                rootDir)
+                    .path + File.separator + "project.properties")
+
+    register("createCerts") {
+        doFirst {
+            if (System.getenv("CI") == true.toString()) {
+                val droneCertFileName = "Drone.cert.pem"
+                val privateKeyFileName = "Drone.private.key.pem"
+                val rootCAFileName = "root-CA.pem"
+                val certFolderPath = mkdir("certs")
+                val droneCert = file(certFolderPath.path + File.separator + droneCertFileName)
+                val privateKey = file(certFolderPath.path + File.separator + privateKeyFileName)
+                val rootCA = file(certFolderPath.path + File.separator + rootCAFileName)
+                val username = if (certFolderPath.path.contains(File.separator + "user-application" + File.separator)) {
+                    "DroneSecurity: User"
+                } else {
+                    "DroneSecurity: Drone"
+                }
+
+                droneCert.appendText(System.getenv("DRONE_CERT"))
+                privateKey.appendText(System.getenv("PRIVATE_KEY"))
+                rootCA.appendText(System.getenv("ROOT_CA"))
+                val endpoint = System.getenv("ENDPOINT")
+
+                val projectProperties = Properties()
+                projectProperties.setProperty("certsFolderPath", certFolderPath.path + File.separator)
+                projectProperties.setProperty("certificateFile", droneCertFileName)
+                projectProperties.setProperty("privateKeyFile", privateKeyFileName)
+                projectProperties.setProperty("certificateAuthorityFile", rootCAFileName)
+                projectProperties.setProperty("endpoint", endpoint)
+                projectProperties.setProperty("clientID", username)
+                val outputStream = propertiesFile.outputStream()
+                projectProperties.store(outputStream, "Connection settings")
+                outputStream.close()
+            } else {
+                copy {
+                    from(rootDir)
+                    include("project.properties")
+                    into(projectDir.path)
+                }
+            }
+        }
+    }
+
+    register("clearCerts") {
+        doLast {
+            if (System.getenv("CI") == true.toString()) {
+                val certFolderPath = file(projectDir.path + File.separator + "certs")
+                certFolderPath.deleteRecursively()
+            }
+            propertiesFile.delete()
+        }
+    }
+
     register<Jar>("DroneFatJar") {
         group = "build"
         description = "Assembles a runnable fat jar archive containing all the needed stuff to be executed as standalone."
